running

+++ gpu_implementation.cu
@@ -213,7 +213,7 @@ int main() {
         printf("TOOK %f SECONDS TO RUN THIS TEST");
 
         for(int char_num = 0; char_num<100;char_num++){
-            printf("%d",int(output_message[char_num]));
+            printf("%d",(int)output_message[char_num]);
         }
         // printf(output_message);
 
@@ -221,6 +221,6 @@ int main() {
         total_time+=test_time;
     }
     printf("Done.\n");
-    printf("AVERAGE TIME ACROSS %d TESTS: %f ms\n",NUM_TESTS, ceilf(float(total_time)/float(NUM_TESTS)));
+    printf("AVERAGE TIME ACROSS %d TESTS: %f ms\n",NUM_TESTS, (double)(ceilf(float(total_time)/float(NUM_TESTS)));
     return 0;
 }
